project-elsd-dsl-setup/requirements.txt
flask==3.0.2
antlr4-python3-runtime==4.13.2
matplotlib==3.8.3
pillow==10.2.0 